pyproject.toml
[project]
name = "air-quality-backend"
version = "0.1.0"
description = "室内空气质量检测系统后端API - Django + Django REST Framework"
readme = "README.md"
requires-python = ">=3.12.0"
authors = [
    { name = "Air Quality Team", email = "team@airquality.com" }
]
keywords = ["django", "air-quality", "detection", "ocr", "reports"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    # Django核心
    "django>=4.2.0,<5.0.0",
    "djangorestframework>=3.14.0",
    # 数据库
    "psycopg2-binary>=2.9.0",
    # API文档
    "drf-spectacular>=0.26.0",
    # CORS支持
    "django-cors-headers>=4.0.0",
    # 异步任务队列
    "celery>=5.3.0",
    "redis>=4.5.0",
    # 环境变量
    "python-dotenv>=1.0.0",
    # 图片处理
    "pillow>=10.0.0",
    # Excel处理
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    # Word文档处理
    "python-docx>=0.8.11",
    # PDF处理
    "reportlab>=4.0.0",
    # HTTP请求
    "requests>=2.31.0",
    # 时间处理
    "python-dateutil>=2.8.0",
    # 文件类型检测
    "python-magic>=0.4.27",
    # 字符串相似度计算
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.21.0",
    # 数据验证
    "jsonschema>=4.17.0",
    # 监控和日志
    "sentry-sdk>=1.32.0",
    # AI/OCR相关
    "google-generativeai>=0.8.5",
    "openai>=1.90.0",
    "pygithub>=2.6.1",
    "docx2pdf>=0.1.8",
]

[project.optional-dependencies]
dev = [
    # 开发工具
    "django-debug-toolbar>=4.1.0",
    "django-extensions>=3.2.0",

    # 代码质量
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "django-stubs>=4.2.0",

    # 预提交钩子
    "pre-commit>=3.3.0",
]

test = [
    # 测试框架
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",

    # 测试数据
    "factory-boy>=3.3.0",
    "faker>=19.0.0",

    # 模拟和断言
    "responses>=0.23.0",
    "freezegun>=1.2.0",
]

docs = [
    # 文档生成
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]

production = [
    # 生产环境
    "gunicorn>=21.0.0",
    "whitenoise>=6.5.0",
    "django-storages>=1.14.0",
    "boto3>=1.28.0",  # AWS S3支持
]

all = [
    "air-quality-backend[dev,test,docs,production]"
]

[project.urls]
Homepage = "https://github.com/your-org/air-quality-backend"
Documentation = "https://air-quality-backend.readthedocs.io/"
Repository = "https://github.com/your-org/air-quality-backend.git"
"Bug Tracker" = "https://github.com/your-org/air-quality-backend/issues"

[project.scripts]
air-quality-manage = "manage:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "air-quality-backend[dev,test]"
]

[tool.uv.sources]
# 如果需要从特定源安装包，可以在这里配置

[tool.hatch.build.targets.wheel]
packages = ["apps"]

[tool.hatch.metadata]
allow-direct-references = true

# Black代码格式化配置
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

# isort导入排序配置
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = ["apps", "config"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = ["*/migrations/*"]

# MyPy类型检查配置
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

# Pytest配置
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*", "*Tests", "*TestCase"]
python_functions = ["test_*"]
testpaths = ["apps", "tests"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=apps",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "api: marks tests as API tests",
    "ocr: marks tests related to OCR functionality",
    "reports: marks tests related to report generation",
    "monthly: marks tests related to monthly reports",
    "batch: marks tests related to batch processing",
    "core: marks tests related to core services",
    "files: marks tests related to file management",
]

# Coverage配置
[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "config/wsgi.py",
    "config/asgi.py",
    "*/tests.py",
    "*/test_*.py",
    "*/*_test.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
